Connector methods, tests, and fixes
