+++ citadel-internal-service-connector/Cargo.toml
@@ -14,5 +14,4 @@ bincode2 = { workspace = true }
 serde = { workspace = true }
 futures = { workspace = true, features = ["alloc"] }
 uuid = { workspace =  true, features = ["v4"]}
-citadel_logging = { workspace = true }
 async-trait = "0.1.79"
