+++ citadel-internal-service-connector/src/lib.rs
@@ -1,6 +1,3 @@
 pub mod codec;
 pub mod connector;
-
 pub mod io_interface;
-pub mod connector;
-pub mod util;
