+++ citadel-internal-service/tests/utilities.rs
@@ -4,7 +4,11 @@ mod common;
 mod tests {
     use crate::common::server_info_skip_cert_verification;
     use citadel_internal_service::kernel::CitadelWorkspaceService;
-    use citadel_internal_service_connector::connector::InternalServiceConnector;
+    use citadel_internal_service_connector::connector::{ClientError, InternalServiceConnector};
+    use citadel_internal_service_connector::io_interface::tcp::TcpIOInterface;
+    use citadel_internal_service_connector::scan_for_response;
+    use citadel_internal_service_types::InternalServiceResponse::PeerConnectSuccess;
+    use citadel_internal_service_types::{ConnectSuccess, InternalServiceResponse};
     use citadel_sdk::prelude::{BackendType, NodeBuilder, NodeType, SecBuffer};
     use std::error::Error;
     use std::net::SocketAddr;
@@ -35,7 +39,7 @@ mod tests {
 
         // Start Internal Service
         let internal_service_kernel =
-            CitadelWorkspaceService::new(SocketAddr::from_str("127.0.0.1:23457")?);
+            CitadelWorkspaceService::new_tcp(SocketAddr::from_str("127.0.0.1:23457")?).await?;
         let internal_service = NodeBuilder::default()
             .with_node_type(NodeType::Peer)
             .with_backend(BackendType::InMemory)
@@ -67,7 +71,7 @@ mod tests {
         crate::common::setup_log();
         let (server, server_bind_address) = server_info_skip_cert_verification();
         tokio::task::spawn(server);
-        let _service_connector_0 = connector_service_and_server(
+        let (mut service_connector_0, cid_0) = connector_service_and_server(
             server_bind_address,
             SocketAddr::from_str("127.0.0.1:23457")?,
             "name 0",
@@ -75,7 +79,7 @@ mod tests {
             "password0",
         )
         .await?;
-        let _service_connector_1 = connector_service_and_server(
+        let (mut service_connector_1, cid_1) = connector_service_and_server(
             server_bind_address,
             SocketAddr::from_str("127.0.0.1:23458")?,
             "name 1",
@@ -83,7 +87,30 @@ mod tests {
             "password1",
         )
         .await?;
-        // TODO: Add Peer Register and Connect - Requires CID from Responses above
+
+        let peer_0_register_and_connect = tokio::task::spawn(async move {
+            service_connector_0
+                .peer_register_and_connect_with_defaults(cid_0, cid_1)
+                .await
+        });
+        let peer_1_register_and_connect = tokio::task::spawn(async move {
+            service_connector_1
+                .peer_register_and_connect_with_defaults(cid_1, cid_0)
+                .await
+        });
+        let result =
+            futures::future::join_all([peer_0_register_and_connect, peer_1_register_and_connect])
+                .await;
+        if result.iter().all(|i| i.is_ok()) {
+            citadel_logging::info!(target: "citadel", "Peers Registration and Connection to each other was successful");
+        } else {
+            panic!("Peer Register and Connect Error")
+        }
+        // service_connector_0.peer_register_with_defaults(cid_0, cid_1).await?;
+        // service_connector_1.peer_register_with_defaults(cid_1, cid_0).await?;
+        //
+        // service_connector_0.peer_connect_with_defaults(cid_0, cid_1).await?;
+        // service_connector_1.peer_connect_with_defaults(cid_1, cid_0).await?;
         Ok(())
     }
 
@@ -93,8 +120,8 @@ mod tests {
         full_name: S,
         username: S,
         password: R,
-    ) -> Result<InternalServiceConnector, Box<dyn Error>> {
-        let internal_service_kernel = CitadelWorkspaceService::new(service_addr);
+    ) -> Result<(InternalServiceConnector<TcpIOInterface>, u64), Box<dyn Error>> {
+        let internal_service_kernel = CitadelWorkspaceService::new_tcp(service_addr).await?;
         let internal_service = NodeBuilder::default()
             .with_node_type(NodeType::Peer)
             .with_backend(BackendType::InMemory)
@@ -106,7 +133,7 @@ mod tests {
         // Connect to Internal Service via TCP
         let mut service_connector =
             InternalServiceConnector::connect_to_service(service_addr).await?;
-        service_connector
+        match service_connector
             .register_and_connect(
                 server_addr,
                 full_name,
@@ -114,7 +141,10 @@ mod tests {
                 password,
                 Default::default(),
             )
-            .await?;
-        Ok(service_connector)
+            .await
+        {
+            Ok(ConnectSuccess { cid, request_id: _ }) => Ok((service_connector, cid)),
+            Err(err) => Err(Box::from(err)),
+        }
     }
 }
